+++ htdocs/js/init.js
@@ -292,12 +292,10 @@ function appInit() {
     }
     //update state on window focus - browser pauses javascript
     window.addEventListener('focus', function() {
-        logDebug('Browser tab gots the focus -> update player state');
-        getState();
-        if (app.id === 'QueueCurrent') {
-            execSearchExpression(elGetById('QueueCurrentSearchStr').value);
-        }
-        websocketKeepAlive();
+        onShow();
+    }, false);
+    window.addEventListener('pageshow', function() {
+        onShow();
     }, false);
     //global keymap
     document.addEventListener('keydown', function(event) {
@@ -373,6 +371,19 @@ function appInit() {
     window.addEventListener('beforeunload', function() {
         webSocketClose();
     });
+}
+
+/**
+ * Checks the connection state and reconnects the websocket on demand
+ * @returns {void}
+ */
+function onShow() {
+    logDebug('Browser focused, update player state');
+    getState();
+    if (app.id === 'QueueCurrent') {
+        execSearchExpression(elGetById('QueueCurrentSearchStr').value);
+    }
+    websocketKeepAlive();
 }
 
 /**